requirements.txt
bert-lid==0.0.1
boto3==1.14.43
botocore==1.17.43
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
future==0.18.2
idna==2.10
jmespath==0.10.0
joblib==0.16.0
locationiq==1.1.0
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
protobuf==3.13.0
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-pretrained-bert-lid==0.6.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.0
six==1.15.0
sklearn==0.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0+cu101
tqdm==4.48.2
tweepy==3.9.0
typechecker==0.0.1
Unidecode==1.1.1
urllib3==1.25.10
